move utilities function into its own folder;
